requirements.txt
Pillow
python-osc
